@@ -3,9 +3,9 @@
   "version": "1.0.0",
   "description": "使用React、Redux、Immutable编写「俄罗斯方块」。Use Tetact, Redux, Immutable to coding \"Tetris\".",
   "scripts": {
-    "start": "webpack-dev-server --progress",
-    "build": "rm -rf ./build/* && NODE_ENV=production webpack --config ./webpack.production.config.js --progress && ls ./build",
-    "windowsBuild": "rm -rf ./build/* && set NODE_ENV=production && webpack --config ./webpack.production.config.js --progress && ls ./build"
+    "start": "eslint --ignore-path .gitignore --fix . && webpack-dev-server --progress --open ",
+    "build": "rm -rf ./build/* && NODE_ENV=production && webpack -p --config ./webpack.config.production.babel.js --progress && ls ./build",
+    "windowsBuild": "rm -rf ./build/* && NODE_ENV=production && webpack -p --config ./webpack.config.production.babel.js --progress && ls ./build"
   },
   "repository": {
     "type": "git",
@@ -25,38 +25,39 @@
   },
   "homepage": "https://github.com/chvin/react-tetris#readme",
   "devDependencies": {
-    "babel-core": "^6.13.2",
-    "babel-loader": "^6.2.4",
+    "babel-core": "^6.22.1",
+    "babel-eslint": "^6.1.2",
+    "babel-loader": "^6.2.10",
     "babel-plugin-react-transform": "^2.0.2",
-    "babel-preset-es2015": "^6.13.2",
-    "babel-preset-react": "^6.16.0",
-    "copy-webpack-plugin": "^3.0.1",
-    "css-loader": "^0.23.1",
-    "eslint": "^3.3.1",
-    "eslint-config-airbnb": "^10.0.1",
+    "babel-preset-es2015": "^6.22.0",
+    "babel-preset-react": "^6.22.0",
+    "copy-webpack-plugin": "^4.0.1",
+    "css-loader": "^0.26.1",
+    "eslint": "^3.15.0",
+    "eslint-config-airbnb": "^14.0.0",
     "eslint-loader": "^1.6.1",
-    "eslint-plugin-import": "^1.13.0",
-    "eslint-plugin-jsx-a11y": "^2.1.0",
-    "eslint-plugin-react": "^6.1.1",
-    "extract-text-webpack-plugin": "^1.0.1",
-    "file-loader": "^0.9.0",
-    "html-webpack-plugin": "^2.22.0",
+    "eslint-plugin-import": "^2.2.0",
+    "eslint-plugin-jsx-a11y": "^3.0.2",
+    "eslint-plugin-react": "^6.9.0",
+    "extract-text-webpack-plugin": "^2.0.0-rc.3",
+    "file-loader": "^0.10.0",
+    "html-webpack-plugin": "^2.28.0",
     "json-loader": "^0.5.4",
-    "less": "^2.7.1",
+    "less": "^2.7.2",
     "less-loader": "^2.2.3",
     "react-transform-hmr": "^1.0.4",
     "style-loader": "^0.13.1",
     "url-loader": "^0.5.7",
-    "webpack": "^1.13.1",
-    "webpack-dev-server": "^1.14.1"
+    "webpack-dev-server": "^2.3.0"
   },
   "dependencies": {
     "classnames": "^2.2.5",
     "immutable": "^3.8.1",
-    "react": "^15.3.0",
-    "react-dom": "^15.3.0",
-    "react-redux": "^4.4.5",
-    "redux": "^3.5.2",
-    "redux-immutable": "^3.0.8"
+    "react": "^15.4.2",
+    "react-dom": "^15.4.2",
+    "react-redux": "^5.0.2",
+    "redux": "^3.6.0",
+    "redux-immutable": "^3.0.11",
+    "webpack": "^2.2.1"
   }
 }
