@@ -14,7 +14,7 @@ export default class Music extends React.Component {
             bg: true,
             [style.music]: true,
             [style.c]: !this.props.data,
-          }
+          },
         )}
       />
     );
