@@ -13,6 +13,6 @@ render(
   <Provider store={store}>
     <App />
   </Provider>
-    , document.getElementById('root')
+    , document.getElementById('root'),
 );
 
