@@ -53,7 +53,7 @@ const blankLine = [0, 0, 0, 0, 0, 0, 0, 0, 0, 0];
 
 const blankMatrix = (() => {
   const matrix = [];
-  for (let i = 0; i < 20; i++) {
+  for (let i = 0; i < 20; i += 1) {
     matrix.push(List(blankLine));
   }
   return List(matrix);
@@ -88,7 +88,7 @@ const maxPoint = 999999;
 const transform = (function () {
   const trans = ['transform', 'webkitTransform', 'msTransform', 'mozTransform', 'oTransform'];
   const body = document.body;
-  return trans.filter((e) => body.style[e] !== undefined)[0];
+  return trans.filter(e => body.style[e] !== undefined)[0];
 }());
 
 const eachLines = 20; // 每消除eachLines行, 增加速度
